Add files via upload
This is an exercise which shows some available tools in R to test if a linear model, created from an open-source database, is correctly specified and to check if it accounts for many of the technical hurdles that could affect it. The Document is written in Spanish (my native language).